Rename files to use camel case
