+++ server/middlewares/ensureAuth.js
@@ -1,4 +1,4 @@
-const jwt = require('../libraries/json-web-token');
+const jwt = require('../libraries/jsonWebToken');
 
 /**
  * Ensure Auth Middleware
